#9: Support documentation for built-in commands and variables
