+++ .github/workflows/build_plugin.yml
@@ -1,6 +1,6 @@
 name: Build Plugin
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
   # Run Gradle Wrapper Validation Action to verify the wrapper's checksum
