+++ src/main/kotlin/de/nordgedanken/auto_hotkey/ide/documentation/AhkDocumentationProvider.kt
@@ -1,22 +1,16 @@
 package de.nordgedanken.auto_hotkey.ide.documentation
 
-import com.github.b3er.idea.plugins.arc.browser.formats.SevenZipArchiveFileSystemImpl
 import com.intellij.ide.BrowserUtil
 import com.intellij.lang.documentation.DocumentationProvider
 import com.intellij.lang.documentation.ExternalDocumentationHandler
 import com.intellij.openapi.editor.Editor
-import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.io.FileTooBigException
-import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.impl.ArchiveHandler
-import com.intellij.openapi.vfs.newvfs.impl.StubVirtualFile
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import com.intellij.psi.impl.source.tree.LeafPsiElement
 import de.nordgedanken.auto_hotkey.lang.psi.AhkTokenType
-import de.nordgedanken.auto_hotkey.project.settings.defaultAhkSdk
-import de.nordgedanken.auto_hotkey.util.AhkBundle
 import java.io.FileNotFoundException
 
 /**
@@ -43,7 +37,7 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         element ?: return null
 
         val chm = try {
-            getChmArchiveHandler(element.project)
+            ChmArchiveUtil.getChmArchiveHandler(element.project)
         } catch (e: IllegalStateException) {
             return null
         }
@@ -51,7 +45,7 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         if (element.text.startsWith("A_"))
             return "https://www.autohotkey.com/docs/Variables.htm#" + element.text.drop(2)
 
-        val pathInChm = getPathInChm(chm, element.text)
+        val pathInChm = ChmArchiveUtil.getPathInChm(chm, element.text)
             ?: return null
 
         return "https://www.autohotkey.com/$pathInChm"
@@ -62,7 +56,7 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         if (element.elementType !is AhkTokenType) return null
 
         val chm = try {
-            getChmArchiveHandler(element.project)
+            ChmArchiveUtil.getChmArchiveHandler(element.project)
         } catch (e: IllegalStateException) {
             return e.message
         }
@@ -70,14 +64,14 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         if (element.text.startsWith("A_"))
             return extractVariableText(chm, element.text)
 
-        val pathInChm = getPathInChm(chm, element.text)
+        val pathInChm = ChmArchiveUtil.getPathInChm(chm, element.text)
             ?: return null
 
         return extractHtmlText(chm, pathInChm)
     }
 
     private fun extractVariableText(chm: ArchiveHandler, variableName: String): String? {
-        val variablesPage = getPathInChm(chm, "Variables")
+        val variablesPage = ChmArchiveUtil.getPathInChm(chm, "Variables")
             ?: return null
 
         return try {
@@ -111,49 +105,6 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         }
     }
 
-    private fun getPathInChm(
-        chm: ArchiveHandler,
-        approximateTitle: String?
-    ): String? {
-        val paths = arrayOf(
-            "docs/commands",
-            "docs/misc",
-            "docs/objects",
-            "docs",
-            "docs/scripts"
-        )
-
-        for (path in paths) {
-            val files = chm.list(path)
-            for (fileName in files) {
-                if (fileName.equals("$approximateTitle.htm", true) ||
-                    fileName.equals("_$approximateTitle.htm", true)
-                )
-                    return "$path/$fileName"
-            }
-        }
-        return null
-    }
-
-    private fun getChmArchiveHandler(project: Project): ArchiveHandler {
-        val ahkSdk = project.defaultAhkSdk
-            ?: error(AhkBundle.msg("documentation.error.no.runner.configured"))
-
-        val homeDirectory: VirtualFile = ahkSdk.homeDirectory
-            ?: error(AhkBundle.msg("documentation.error.cannot.access.runner.home.dir"))
-
-        val chmFile: VirtualFile = homeDirectory.findFileByRelativePath("AutoHotkey.chm")
-            ?: error(AhkBundle.msg("documentation.error.chm.file.not.found.in.home.dir"))
-
-        val stub = object : StubVirtualFile() {
-            override fun getPath(): String = "${chmFile.path}!/"
-            override fun getParent(): VirtualFile? = null
-        }
-
-        return SevenZipArchiveFileSystemImpl.instance.getHandlerForFile(stub)
-            ?: error("Error initializing 7zip file system")
-    }
-
     override fun handleExternal(
         element: PsiElement?,
         originalElement: PsiElement?
@@ -182,7 +133,7 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
         element ?: return "Cannot fetch documentation"
 
         val chm = try {
-            getChmArchiveHandler(element.project)
+            ChmArchiveUtil.getChmArchiveHandler(element.project)
         } catch (e: IllegalStateException) {
             return e.message ?: "Error fetching documentation"
         }
@@ -191,7 +142,7 @@ class AhkDocumentationProvider : DocumentationProvider, ExternalDocumentationHan
             .substringBefore('#')
             .substringBeforeLast('.')
 
-        val pathInChm = getPathInChm(chm, page)
+        val pathInChm = ChmArchiveUtil.getPathInChm(chm, page)
             ?: return "Cannot find file in chm file for $page"
 
         return extractHtmlText(chm, pathInChm)
